--performFullCheck forces Laputin to check hashes even though files are in same path as before
This was needed due to a bug which created incorrect hashes for files and because those files' paths were up to date Laputin did not correct the mistake after hashing had been fixed.

This also fixes quick check to check file size in addition to file path when skipping hashing to improve reliability.