Fixed tests with slight delay
File changes are not signaled to tests at all so they expect file changes to be correct when file library emits an event. There might be some delay between that and actual file changes however.

To fix these properly Laputin should probably emit some signal when file changes have actually been performed but that is outside the scope of this commit.