Fix QuickMD5Hasher to work with files with mostly different content but identical middle chunk
This is a bit ugly but the idea is to find a chunk of data which contains non-identical bytes. In practice files with different content should contain non-identical middle chunk but in practice a couple of cases were observed where a middle chunk (and few chunks next to it) were full of uint 255 values.